feat: post 테스트 추가

+++ src/main/kotlin/kr/pe/karsei/boardtraffic/adapter/out/PostJpaAdapter.kt
@@ -37,6 +37,8 @@ class PostJpaAdapter(
     }
 
     override fun deletePost(postId: Long) {
+        val exists = postRepository.existsById(postId)
+        if (!exists) throw ResponseStatusException(HttpStatus.NOT_FOUND, "포스트가 존재하지 않습니다.")
         postRepository.deleteById(postId)
     }
 }
